add class FormatterUtil and addition method shouldLoadFromFile()
