+++ src/managers/FileBackedTaskManager.java
@@ -1,7 +1,6 @@
 package managers;
 import tasks.*;
 import java.io.*;
-import java.util.ArrayList;
 import java.util.List;
 
 public class FileBackedTaskManager extends InMemoryTaskManager {
@@ -14,88 +13,30 @@ public class FileBackedTaskManager extends InMemoryTaskManager {
         this.file = file;
     }
     public void save() {
-        try {
             try (BufferedWriter bufferedWriter = new BufferedWriter(new FileWriter(file))) {
                 bufferedWriter.write(tasksAndHistoryToString());
             } catch (IOException e) {
                 throw new ManagerSaveException("Ошибка сохранения");
             }
-        } catch (ManagerSaveException e) {
-            System.out.println(e.getMessage());
-        }
     }
 
-    public String taskToString(Task task){
-        TaskType taskType = TaskType.TASK;
-        int epicId = -1;
-        if (task instanceof Epic) {
-            taskType = TaskType.EPIC;
-        } else if (task instanceof SubTask) {
-            taskType = TaskType.SUBTASK;
-            epicId = ((SubTask) task).getEpicId();
-        }
-        return task.getId() + "," + taskType + "," + task.getName() + "," + task.getTaskStatus() + ","
-                + task.getDescription() + "," + epicId;
-    }
     public String tasksAndHistoryToString(){
         StringBuilder builder = new StringBuilder("id,type,name,status,description,epic\n");
 
         for (Task task : tasks.values()) {
-            builder.append(taskToString(task));
+            builder.append(FormatterUtil.taskToString(task));
             builder.append("\n");
         }
         for (Epic epic : epics.values()) {
-            builder.append(taskToString(epic));
+            builder.append(FormatterUtil.taskToString(epic));
             builder.append("\n");
         }
         for (SubTask subTask : subTasks.values()) {
-            builder.append(taskToString(subTask));
+            builder.append(FormatterUtil.taskToString(subTask));
             builder.append("\n");
         }
-        builder.append(historyToString(historyManager));
+        builder.append(FormatterUtil.historyToString(historyManager));
         return builder.toString();
-    }
-
-    public static String historyToString(HistoryManager manager){
-        List<Task> taskList = manager.getHistory();
-        StringBuilder builder = new StringBuilder("History");
-        for(Task task : taskList) {
-            builder.append(",");
-            builder.append(task.getId());
-        }
-        return builder.toString();
-    }
-
-    public Task taskFromString(String value){
-        TaskStatus taskStatus;
-
-        String[] split = value.split(",");
-        if (split[3].equals("NEW")) {
-            taskStatus = TaskStatus.NEW;
-        } else if (split[3].equals("IN_PROGRESS")) {
-            taskStatus = TaskStatus.IN_PROGRESS;
-        } else {
-            taskStatus = TaskStatus.DONE;
-        }
-
-        if (split[1].equals("TASK")) {
-            Task task = new Task(split[2], split[4]);
-            task.setId(Integer.parseInt(split[0]));
-            task.setTaskStatus(taskStatus);
-            return task;
-        } else if (split[1].equals("EPIC")) {
-            Epic epic = new Epic(split[2], split[4]);
-            epic.setId(Integer.parseInt(split[0]));
-            epic.setTaskStatus(taskStatus);
-            return epic;
-        } else if (split[1].equals("SUBTASK")) {
-            SubTask subTask = new SubTask(split[2], split[4]);
-            subTask.setId(Integer.parseInt(split[0]));
-            subTask.setTaskStatus(taskStatus);
-            subTask.setEpicId(Integer.parseInt(split[5]));
-            return subTask;
-        }
-        return null;
     }
 
     public void loadTask(Task task) {
@@ -110,15 +51,6 @@ public class FileBackedTaskManager extends InMemoryTaskManager {
         if (id <= task.getId()) {
             id = task.getId() + 1;
         }
-    }
-
-    static List<Integer> historyFromString(String value){
-        List<Integer> hystoryList = new ArrayList<>();
-        String[] split = value.split(",");
-        for (int i = 1; i < split.length; i++) {
-            hystoryList.add(Integer.parseInt(split[i]));
-        }
-        return hystoryList;
     }
 
     public void loadHistory(List<Integer> hystoryList) {
@@ -139,10 +71,10 @@ public class FileBackedTaskManager extends InMemoryTaskManager {
             while (bufferedReader.ready()) {
                 String line = bufferedReader.readLine();
                 if (line.contains("History")){
-                    List<Integer> idHystoryList = historyFromString(line);
+                    List<Integer> idHystoryList = FormatterUtil.listHistoryFromString(line);
                     loadHistory(idHystoryList);
                 } else if (line.contains("TASK") || line.contains("EPIC") || line.contains("SUBTASK")) {
-                    Task task = taskFromString(line);
+                    Task task = FormatterUtil.taskFromString(line);
                     loadTask(task);
                 }
             }
