+++ src/managers/ManagerSaveException.java
@@ -1,6 +1,6 @@
 package managers;
 
-public class ManagerSaveException extends Exception {
+public class ManagerSaveException extends RuntimeException {
     public ManagerSaveException(String message) {
         super(message);
     }
